* release-v0.0.2: (85 commits)
  Feature list for v0.0.2.
  Disable custom debugging for release.
  Branched to 'release-v0.0.2'.
  Changes to CREATE_ZIP function for CM7's build.prop.
  Disabling automatic stock recovery installation in *.rc files (install-recovery.sh).
  Use header file to determine debugging.
  Manually added new EI22 modules to initramfs_cm7 to stay sync'd with EpicCM.
  Changes to the metadata for the update.zip (I doubt this matters at all).
  Minor changes to ncBuildHelper script.
  Small change to .gitignore.
  Tweaking EXT4 mounts in init.rc for cm7.
  Disabling DEBUG_INFO in cm defconfig.
  Updating cyanogenmod_epic defconfig with all the new features included in this kernel (IKCONFIG, SCHED_AUTOGROUP, schedulers, governors, TUN, Voodoo Sound, etc).
  Enable kexec in cyanogenmod_epic defconfig (kernel would not boot without it).
  Changing scripts to better support building for cyanogenmod and building stand-alone modules.
  Some scripting changes to ncBuildHelper and functions.
  Splitting the functions out of ncBuildHelper and sourcing them in.
  Changes to ncBuildHelper script (fixing OUTFILE, atempting to fix tar.md5 creation for Odin).
  Adjusting reboot flags for compatibility.
  Change reboot recovery to boot bml8 instead of bml7 and changed bml7 reboot mode from recovery to bml7recovery  (based on mkasick's patch)
  ...

+++ initramfs_tw/nubernel/scripts/init_01_system_apps.sh
@@ -23,14 +23,17 @@ REMOVE_SYSTEM_APP()
 	shift
 	local APK="$1"
 	local DATA="$2"
+	SEND_LOG "  test -f /system/app/${APK}.apk"
 	if /sbin/busybox test -f "/system/app/${APK}.apk"; then
 		SEND_LOG "  rm -f /system/app/${APK}.apk"
 		/sbin/busybox rm -f "/system/app/${APK}.apk"
 	fi
+	SEND_LOG "  test -f /system/app/${APK}.odex"
 	if /sbin/busybox test -f "/system/app/${APK}.odex"; then
 		SEND_LOG "  rm -f /system/app/${APK}.odex"
 		/sbin/busybox rm -f "/system/app/${APK}.odex"
 	fi
+	SEND_LOG "  test -d /data/data/${DATA}"
 	if /sbin/busybox test -d "/data/data/${DATA}"; then
 		SEND_LOG "  rm -rf /data/data/${DATA}"
 		/sbin/busybox rm -rf "data/data/${DATA}"
@@ -44,9 +47,14 @@ MOVE_SYSTEM_APP()
 	local APK="$1"
 	local DATA="$2"
 	local CACHE="system@app@${APK}.apk@classes.dex"
+	SEND_LOG "  test -f /system/app/${APK}.odex"
+	if /sbin/busybox test -f "/system/app/${APK}.odex"; then
+		SEND_LOG "  Detected odex file, app will not be moved"
+		return 1
+	fi
 	SEND_LOG "  test -f /system/app/${APK}.apk"
 	if /sbin/busybox test -f "/system/app/${APK}.apk"; then
-		SEND_LOG "  Detected apk file, moving to /data/app"
+		SEND_LOG "  Detected ${APK}.apk file, moving to /data/app"
 		local EXISTING_APP_FIND=$(/sbin/busybox find /data/app -iname "${DATA}*")
 		for EXISTING_APP in $EXISTING_APP_FIND ; do
 			SEND_LOG "  rm -f $EXISTING_APP"
@@ -57,14 +65,6 @@ MOVE_SYSTEM_APP()
 		/sbin/busybox chown system.system "/data/app/${DATA}-1.apk"
 		/sbin/busybox chmod 0644 "/data/app/${DATA}-1.apk"
 
-		SEND_LOG "  test -f /system/app/${APK}.odex"
-		if /sbin/busybox test -f "/system/app/${APK}.odex"; then
-			SEND_LOG "  Detected odex file, moving to /data/app"
-			SEND_LOG "  mv -f /system/app/${APK}.odex /data/app/${DATA}-1.odex"
-			/sbin/busybox mv -f "/system/app/${APK}.odex" "/data/app/${DATA}-1.odex"
-			/sbin/busybox chown system.system "/data/app/${DATA}-1.odex"
-			/sbin/busybox chmod 0644 "/data/app/${DATA}-1.odex"
-		fi
 		SEND_LOG "  test -d /data/data/${DATA}"
 		if /sbin/busybox test -d "/data/data/${DATA}"; then
 			SEND_LOG "  rm -rf /data/data/${DATA}"
